+++ wms_haidh/wms-backend/src/main/java/openerp/openerpresourceserver/projection/DeliveryOrderItemProjection.java
@@ -1,14 +1,13 @@
 package openerp.openerpresourceserver.projection;
 
-import java.math.BigDecimal;
 import java.util.UUID;
 
 public interface DeliveryOrderItemProjection {
 	UUID getAssignedOrderItemId();
 	UUID getOrderId();
     String getProductName();
-    BigDecimal getWeight();
-    Integer getOriginalQuantity();
+    double getWeight();
+    int getOriginalQuantity();
     String getBayCode();
     String getLotId();
 }
